Fixed issue with uploaded file handling in fundraising tests
The tests currently fail on docker.
Using the TemporaryMediaRootMixin originally developped
for the views tests should fix that.
